# SIMD code generator