--- a/src/Automa.jl
+++ b/src/Automa.jl
@@ -2,8 +2,7 @@ module Automa
 
 using DataStructures: DefaultDict
 using Printf: @sprintf
-using SIMD
-using Libdl
+import ScanByte
 
 include("sdict.jl")
 include("sset.jl")
@@ -13,7 +12,6 @@ const Dict = StableDict
 const Set = StableSet
 
 include("byteset.jl")
-include("simd.jl")
 include("re.jl")
 include("precond.jl")
 include("action.jl")