## Introduction
This PR introduces the `:simd` code generator. At this time, it is very similar to the `:goto` code generator. In fact, the only current change from the `:goto`-generator is that it uses SIMD loop unrolling instead of the manual Automa-generated loop unrolling.

In future PRs, I would like to change the large-scale structure of the code generated by the `:simd` generator (see issue #53), if it turns out to be possible and performant. But for now, this change only pertains to SIMD loop unrolling.

## Benefits
Speed. That's it, really. We can use an "empty" format validator (see #55 on a byte array) to most accurately time Automa's parsing, since neither the `Machine`'s attached actions, nor file I/O will influence the result.

The times are given in median runtimes determined using BenchmarkTools.jl:
```
6 MiB FASTA with 60 chars per line:
Before: 2.255 ms
After:  1.283 ms

1.3 GiB 4-line FASTQ with around 100 nt/read:
Before: 368.734 ms
After:  132.312 ms
```
These are, of course, optimal conditions. Even on the extremely simple [biofast FASTQ benchmark](https://github.com/lh3/biofast/tree/master/fqcnt), processing a file is approximately 2.5x slower than processing a byte vector in RAM. So a doubling of speed of the actual Automa parsing is realistically only a 20 % performance increase. Of course, this does mean that we are moving the performance bottleneck somewhere else, such that e.g. TranscodingStreams now have more room for optimization.
Or maybe it turns out it's more than that. We'll have to see for certain when all the tests are in place and the behaviour is verified to be correct.

## Drawbacks of using the `:simd` generator over the `:goto` generator:
Unfortunately, SIMD code generation comes with a few limitations:

* `getbyte` can no longer be a custom function in a `:simd` generator, it has to be `Base.getindex`. I don't think anyone ever used that functionality, though.
* Unlike the `:goto` generator, the generated code is not pure Base code. That means the "consumer" of the code must have `import`'ed or `using` Automa, else they will have a `NameError`
* The code generation uses information from LLVM to determine which CPU instructions are available. If a `:simd` generator is used to generate Julia code on a modern CPU (less than 10 years old x86 CPU), but the Julia code is executed on another CPU (10 years old, or >5 years old AMD, or RISC CPUs), it will likely error catastrophically. This may happen if cached Julia code is used on another computer.

## Drawbacks of merging this code
* Larger code base and more maintenance, obviously.
* Check the `simd.jl` file. It checks the available instructions, but there is, as far as I know, no truly stable and officially supported way of doing this. So it may error sometime in the future. Since this happens at import time, even users not using the `:simd` generator will be affected if this fails on some systems or in the future.
* Ensuring correct behaviour will be harder, since different CPUs will generate different code. It should not be too hard to test, though, since almost all modern CPUs have all instruction sets and can test the code generated for all CPUs.

## To-do list
* ~~Still fails FASTQ parsing, so there are probably a few bugs to be kinked out~~
* ~~Needs way more tests, ideally 100% coverage~~
* Too much code duplication between `:simd` and `:goto` generator. This needs to be factored out (or does it? It depends on how much progress I'll make on reorganizing the `:simd` codegen along the road)

closes issue #54