@jakobnissen 2020-11-07
Note for future: On a FASTA file, the SIMD generator executes 8.7x fewer instructions and 10x fewer branches, while only being 2.5x faster. Part of the reason may be increased cache misses (2.5k -> 4k / 192k), but the major factor is probably the level of instruction level parallism. Looking at the assembly, the SIMD loop has serial memory dependencies, preventing superscalar instructions. On my computer, manually unrolling the SIMD loop does not make it use different registers, so the seriality remains and is in fact slower.
It appears the main bottleneck may be that the CPU simply can't parallelize the vector instructions to the same degree.
It might be worth looking into prefetching though