@kescobo 2020-12-18
> Unlike the :goto generator, the generated code is not pure Base code. That means the "consumer" of the code must have import'ed or using Automa, else they will have a NameError

Might be worth checking how much this increases load time for some consumer package. Automa is pretty light in terms of dependencies, and it sounds like it would be worth it, but it might be nice to have that information at hand.