@jakobnissen 2020-12-18
That's a good idea. The new SIMD.jl dependency has historically been devastating for load times, although this should have improved quite significantly with 1.6, and I don't plan on merging this before then.