@jakobnissen 2020-12-19
Turns out compile times are affected, but not by much. On version 1.5, compiling and running a FASTA reader takes about 2.3 seconds without this PR, and 2.55 seconds with this PR. On 1.6, it's 1.85 seconds without and 1.95 with.