@jakobnissen 2020-12-19
I think this is done now and can get merged. Code coverage is a little misleading, it's actually almost 100%. I've no coverage on the branches that activates on big-endian CPUs and on CPUs without SIMD capabilities, because I don't have those kinds of CPUs. I'ts just simple if-statements though, so not much can go wrong.