@jakobnissen 2020-12-20
Okay, now I've added a whole bunch of comments and also mitigated some of the 1.6 regressions. You're right, it's not very clear what it does. Hopefully the comments help.
Worst case scenario, if, say, in the future this falls apart, the whole SIMD stuff can just be removed, and the SIMD CodeGenContext fall back to returning a goto CodeGenContext.

Although I do think that this whole SIMD code could probably be split out into its own package. I can see how having SIMD byte matching could be generally useful. It is essentially like https://docs.rs/memchr/2.3.4/memchr/, except this generalizes to any collection of bytes. But that's a PR for another time :)