@jakobnissen 2021-03-20
Okay, I've been running on this branch on my daily work for a few months now, refactored it out to use ScanByte, and added non-SIMD fallbacks. So merging now.